backend/requirements.txt
fastapi==0.109.0
sqlmodel>=0.0.22
asyncpg>=0.30.0
pydantic>=2.10.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Authentication dependencies
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.1.3  # Pin to 4.x for passlib compatibility
email-validator>=2.1.0
slowapi>=0.1.9
python-multipart>=0.0.6
